FIX: Updated copyright date and entity to be consistent with License file.
To avoid any confusion regarding the license vs. the repository owner name, updated the
copyright holder to match the license file.

The original entity '3Hour Software' is an internal name from an as-yet-unpublished
project where this code was originally sourced.  For anyone reading the change lists,
'cosaazul' and '3Hour Software' are equivalent for the purposes of ownership of this
copyrighted material.

Details:
* Observable.swift - Updated copyright date & owner

* Observation.swift - Same