FIX (#1, #2): Eliminates strong reference cycles for binding.  Adds closure obsolescence.
Binding closures are no longer methods in the `@Observable` property wrapper, thus use
a capture list to specify a weak reference to `self`.  This allows either side of the
binding to be freed even if `unbind()` is not used.  The remaining half of the binding
will still hold the closure installed by the now-released object, but the associated
`Observation` object is marked obsolete, disabling it and flagging the closure for later
release.

The code adds a new `Observation` property, `isObsolete`, that, once set, disables the
observation and prevents access to its settable attributes (in other words, one can no
longer enable or de-obsolete it).  This blocks access to the closure in the observation
array and triggers it to be removed during subsequent calls to all of the `@Observable`
APIs (`observe`, `remove`, `bind`, `unbind`, and `cleanse`).  For performance reasons,
automatic removal does not happen during the closure callback sequence.  While bindings
rely on the new obsolete state, any observer code can use it to passively mark its
observation for removal, especially when calling `remove()` is impractical (e.g. the
observed property is no longer in the scope of the caller to do so).

A new `@Observable` API, `cleanse()` has been added to facilitate the release of
obsolete observations since there is no guarantee the other APIs will be called.  The
intention is to provide a mechanism for code controlling the observed property to use to
remove dead closures.  A future update may provide a global variant of cleanse() that
will purge all obsolete closures from all active observables in memory.

The documentation in the Readme was updated to reflect the new call and behaviors, as well
clarify some minor points in the original text.  Test code was added for the new states.

This change also simplifies the for-in loop in array_walk(), which was incorrectly crafted
with C++ in mind rather than Swift.

Details:
* Observable.swift -
  1. Added deinit; API: cleanse; func: install_binding;
     & private: clean_obsolete_observations
  2. Removed func: binding_observer
  3. observe() & remove() call clean_obsolete_observations()
  4. bind() calls install_bindiing(), not observe()
  5. array_walk() eliminates C++-style loop range, which is less efficient in Swift

* Observation.swift -
  1. Added computed prop: isObsolete; & private props: _bDead & _kind
  2. Converted prop `kind` to a computed property
  3. Converted explanatory comments to markdown & added text describing isObsolete

* SwiftObservablesTests.swift -
  1. Added funcs: testObservations & testObsolescence; & consts: defaultValue_[x/label]
  2. consts defaultValue_[x/label] used throughout for readability

* Readme.md -
  1. Replaced caveat about binding causing strong reference cycles
  2. Various minor formatting and textual tweaks
  3. Identified "$ syntax" as "projected value" for clarity throughout
  4. Added comments about closure capture and object lifetimes in Basic Usage section
  5. Added note about closure execution order and the new isObsolete functionality
  6. Noted that binding is now a weak relationship in Binding section
  7. Added note to the remove() discussion indicating obsoleting as an alternative
  8. Added section "Obsoleting Closures"